Use thread-local storage in valid()
Before this change, `bench_deref` was 7.52 ns/iter on my machine. After
this change, it's 0.18 ns/iter. I think the difference is that
`thread::current().id()` always clones and drops an `Arc` internally,
while the new version is two non-atomic loads and a predictable branch
in the hot (initialized) path.

As far as I can tell, the only difference between this approach and what
the nightly-only `std::thread::current_id` does, is that the latter is
guaranteed not to allocate, even the first time you call it. I don't
think that's a problem for this crate?